Rename g.s.engine.framework.headlessGameServer -> o.t.game.server
