+++ game-core/src/main/java/games/strategy/engine/framework/startup/launcher/ServerLauncher.java
@@ -15,9 +15,10 @@ import java.util.logging.Level;
 import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
 
+import org.triplea.game.server.HeadlessGameServer;
+
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.framework.ServerGame;
-import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 import games.strategy.engine.framework.message.PlayerListing;
 import games.strategy.engine.framework.startup.mc.ClientModel;
 import games.strategy.engine.framework.startup.mc.GameSelectorModel;
