+++ game-core/src/main/java/games/strategy/triplea/delegate/EndRoundDelegate.java
@@ -11,13 +11,14 @@ import java.util.stream.StreamSupport;
 
 import javax.swing.JOptionPane;
 
+import org.triplea.game.server.HeadlessGameServer;
+
 import games.strategy.engine.data.CompositeChange;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.PlayerList;
 import games.strategy.engine.data.Territory;
 import games.strategy.engine.delegate.IDelegateBridge;
-import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 import games.strategy.engine.message.IRemote;
 import games.strategy.sound.SoundPath;
 import games.strategy.triplea.Constants;
