+++ game-headless/src/main/java/org/triplea/game/server/HeadlessGameRunner.java
@@ -1,7 +1,5 @@
 package org.triplea.game.server;
 
-import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
-
 /**
  * Runs a headless game server.
  */
